# apps/backend/docs/swagger.yaml
openapi: 3.0.0
paths:
...
  /api/queens/{id}:
    get:
      operationId: QueensController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      tags: *ref_4
    patch:
      operationId: QueensController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      tags: *ref_4
    delete:
      operationId: QueensController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      tags: *ref_4
...
info:
  title: API Example
  description: The API description
  version: '1.0'
  contact: {}
...
servers: []
...